move matrix builder
